Splits pager in general and compliance versions
